Split Filter types into separate module
